Adjust constants for increased difficulty and replace doubles with floats for consistency

+++ include/constants.h
@@ -4,7 +4,7 @@
 #include <iostream>
 using namespace std;
 
-const double SCREEN_WIDTH = 1000;
-const double SCREEN_HEIGHT = 625;
-const double GROUND_Y = SCREEN_HEIGHT - 100;
-const double OBSTACLE_LENGTH = 50;
+const float SCREEN_WIDTH = 1000;
+const float SCREEN_HEIGHT = 625;
+const float GROUND_Y = SCREEN_HEIGHT - 100;
+const float OBSTACLE_LENGTH = 50;
